requirements.txt
faiss-cpu
dotenv
openai